유튜브 영상 올린 지점

+++ Assets/02.Scripts/All Inventory/MerChant Inventory/MerChantInventoryManager.cs
@@ -437,7 +437,7 @@ public class MerChantInventoryManager : BaseItemInvenManager
         }
         else
         {
-            possibleBuyAmount = playerRubyAmount / itemPrice;
+            possibleBuyAmount = (int)(playerRubyAmount / itemPrice);
             return true;
         }
     }
